# openapi.yaml
openapi: 3.0.3
info:
  title: ias_stt — Trap Management API
  version: 1.0.0
  description: |
    Trap management service with deployment tracking, location
    history, and photo upload. Telemetry is ingested from the
    ChirpStack HTTP integration via `POST /api/telemetry/ingest`
    (an optional MQTT listener remains available). Application users obtain
    JWT access and refresh tokens from `/auth/login` and send access tokens in
    the `Authorization` header. A legacy API key remains available for service
    integrations.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/deployments:
    get:
      tags: [Deployments]
      summary: List deployments
      operationId: listDeployments
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: trap_id
          in: query
          schema: { type: integer }
          description: Filter by trap ID
        - name: status
          in: query
          schema: { type: string, enum: [active, closed] }
          description: Filter by deployment status
      responses:
        200:
          description: A JSON array of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'

    post:
      tags: [Deployments]
      summary: Manually create a deployment
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentCreate'
            example:
              trap_id: 1
              location: north
              notes: Initial placement
      responses:
        201:
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
  /api/deployments/{deployment_id}:
    parameters:
      - name: deployment_id
        in: path
        required: true
        schema: { type: integer }
        description: Deployment primary key ID

    get:
      tags: [Deployments]
      summary: Get a deployment by ID
      operationId: getDeployment
      responses:
        200:
          description: Deployment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'

    put:
      tags: [Deployments]
      summary: Update deployment details
      operationId: updateDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentUpdate'
            example:
              animal_capture: 2 rats
              notes: Updated notes
      responses:
        200:
          description: Deployment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'

    delete:
      tags: [Deployments]
      summary: Delete a deployment
      operationId: deleteDeployment
      responses:
        200:
          description: Deployment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
              example:
                message: Deployment 1 deleted
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
  /api/deployments/{deployment_id}/photo:
    parameters:
      - name: deployment_id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Deployments]
      summary: Upload a photo for a deployment
      operationId: uploadPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (jpg, jpeg, png, gif)
              required: [file]
      responses:
        200:
          description: Photo uploaded; `photo_url` and `photo_filename` are set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
  /api/deployments/{deployment_id}/photos/{photo_id}:
    parameters:
      - name: deployment_id
        in: path
        required: true
        schema: { type: integer }
      - name: photo_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Deployments]
      summary: Retrieve an authenticated deployment photo
      operationId: getDeploymentPhoto
      responses:
        200:
          description: Image bytes
          content:
            image/jpeg: {}
            image/png: {}
            image/gif: {}
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
  /api/deployments/{deployment_id}/locations:
    parameters:
      - name: deployment_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Deployments]
      summary: List location history for a deployment
      operationId: deploymentLocations
      responses:
        200:
          description: Location history array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeploymentLocation'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
  /api/deployments/{deployment_id}/locations/latest:
    parameters:
      - name: deployment_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Deployments]
      summary: Get the most recent location for a deployment
      operationId: deploymentLatestLocation
      responses:
        200:
          description: Latest location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentLocation'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'

  # -- Trap-scoped nested ----------------------------------------------------
  /api/traps/{trap_id}/deployments:
    parameters:
      - name: trap_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Deployments]
      summary: List all deployments for a trap
      operationId: trapDeployments
      responses:
        200:
          description: Deployments array (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
  /api/traps/{trap_id}/deployments/active:
    parameters:
      - name: trap_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Deployments]
      summary: Get the active deployment for a trap
      operationId: trapActiveDeployment
      responses:
        200:
          description: Active deployment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'
  /api/traps/{trap_id}/locations:
    parameters:
      - name: trap_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Deployments]
      summary: List all location history across all deployments for a trap
      operationId: trapLocations
      responses:
        200:
          description: Location history array (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeploymentLocation'
        401:
          $ref: '#/components/responses/401'
        404:
          $ref: '#/components/responses/404'

# ===========================================================================
# Components
# ===========================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Application users pass the JWT access token as
        `Authorization: Bearer <access-token>`. Legacy service integrations
        may pass the configured API key in the same header.

  # -- Reusable parameters ---------------------------------------------------
  parameters:
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 100, minimum: 0 }
      description: Maximum items per page
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
      description: Number of items to skip

  # -- Schemas ---------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code: { type: string }
                message: { type: string }
      example:
        error:
          code: AUTHENTICATION_REQUIRED
          message: Authentication is required.
...
    Deployment:
      type: object
      properties:
        id:                { type: integer, readOnly: true }
        trap_id:           { type: integer }
        status:            { type: string, enum: [active, closed] }
        start_date:        { type: string, format: date-time }
        end_date:          { type: string, format: date-time, nullable: true }
        animal_capture:    { type: string, maxLength: 255, nullable: true }
        photo_url:         { type: string, maxLength: 500, nullable: true }
        photo_filename:    { type: string, maxLength: 255, nullable: true }
        notes:
          type: array
          items: { $ref: '#/components/schemas/DeploymentNote' }
        pictures:
          type: array
          items: { $ref: '#/components/schemas/DeploymentPicture' }
        created_at:        { type: string, format: date-time, readOnly: true }
        updated_at:        { type: string, format: date-time, readOnly: true }
    DeploymentCreate:
      type: object
      properties:
        trap_id:           { type: integer }
        location:          { type: string, maxLength: 50 }
        notes:             { type: string, maxLength: 500 }
      required: [trap_id]
    DeploymentUpdate:
      type: object
      properties:
        animal_capture:    { type: string, maxLength: 255 }
        notes:             { type: string, maxLength: 500 }
    DeploymentNote:
      type: object
      properties:
        id: { type: integer }
        deployment_id: { type: integer }
        notes: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DeploymentPicture:
      type: object
      properties:
        id: { type: integer }
        deployment_id: { type: integer }
        photo_url: { type: string }
        photo_filename: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DeploymentLocation:
      type: object
      properties:
        id:                { type: integer, readOnly: true }
        deployment_id:     { type: integer, readOnly: true }
        location:          { type: string, maxLength: 50 }
        recorded_at:       { type: string, format: date-time }
        notes:             { type: string, maxLength: 255, nullable: true }
        created_at:        { type: string, format: date-time, readOnly: true }

  # -- Reusable responses ----------------------------------------------------
  responses:
    400:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingField:
              summary: Missing required field
              value: { error: "Missing required field(s): status" }
            invalidType:
              summary: Invalid limit/offset
              value: { error: "'limit' and 'offset' must be integers" }
    401:
      description: Missing or invalid JWT or service credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required.
...
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Trap not found
...